rename util_classes -> cli_mixins
