+++ src/ophys_etl/modules/video/cli_mixins.py
