+++ src/ophys_etl/modules/video/side_by_side_video.py
@@ -1,4 +1,4 @@
-from ophys_etl.modules.video.util_classes import VideoModuleMixin
+from ophys_etl.modules.video.cli_mixins import VideoModuleMixin
 
 from ophys_etl.modules.video.utils import (
     create_side_by_side_video)
